+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 <!-- next-header -->
 
 ## [Unreleased] - ReleaseDate
+
+## [0.11.0] - 2022-01-09
 ### Added
 - Possible to undo importing an image.
 - Grab tool for creating a brush from a part of an image, and Char Brush tool for drawing with it.
@@ -71,7 +73,8 @@ This makes it impossible to try to draw with colors 8-15 unless in a multicolor 
 
 
 <!-- next-url -->
-[Unreleased]: https://github.com/vilcans/pixel_pen/compare/v0.10.0...HEAD
+[Unreleased]: https://github.com/vilcans/pixel_pen/compare/v0.11.0...HEAD
+[0.11.0]: https://github.com/vilcans/pixel_pen/compare/v0.10.0...v0.11.0
 [0.10.0]: https://github.com/vilcans/pixel_pen/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/vilcans/pixel_pen/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/vilcans/pixel_pen/compare/v0.7.0...v0.8.0